appendix/pyproject.toml
[project]
name = "grading-stats-visualization"
version = "1.0.0"
authors = [{ name = "Michal Šoltis", email = "xsoltis1@fi.muni.cz" }]
dependencies = [
    "dash==2.14.1",
    "dash-mantine-components==0.12.1",
    "gunicorn==21.2.0",
    "pandas==2.1.3",
    "pendulum==2.1.2",
    "plotly==5.18.0",
    "polars==0.19.15",
    "psycopg2-binary==2.9.9",
    "redis==5.0.1",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["src"]

[tool.ruff]
select = ["ALL"]
ignore = [
    "ANN101", # Missing type annotation for `self` in method
    "ANN002", # Missing type annotation for `*args`
    "ANN003", # Missing type annotation for `**kwargs`
    "D100",   # Missing docstring in public module
    "D104",   # Missing docstring in public package
    "D203",   # 1 blank line required before class docstring
    "D213",   # Multi-line docstring summary should start at the second line
]

[tool.mypy]
ignore_missing_imports = true
